+++ activerecord/lib/active_record/encryption/context_for_attribute.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module ActiveRecord
+  module Encryption
+    class ContextForAttribute < Context
+      def merge(other) # :nodoc:
+        assign_properties(other.non_defaults)
+      end
+    end
+  end
+end